refactor(coin-list): Merge CoinListRoute into CoinListScreen
