--- coin-list/presentation/src/main/java/io/soma/cryptobook/coinlist/presentation/CoinListRoute.kt
@@ -1,42 +1,0 @@
-package io.soma.cryptobook.coinlist.presentation
-
-import android.widget.Toast
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import io.soma.cryptobook.core.domain.navigation.AppPage
-import io.soma.cryptobook.core.presentation.collectWithLifecycle
-
-@Composable
-fun CoinListRoute(
-    modifier: Modifier = Modifier,
-    viewModel: CoinListViewModel = hiltViewModel(),
-) {
-    val uiState by viewModel.state.collectAsStateWithLifecycle()
-    val context = LocalContext.current
-
-    viewModel.sideEffect.collectWithLifecycle { effect ->
-        when (effect) {
-            is CoinListSideEffect.NavigateToCoinDetail -> viewModel.navigationHelper.navigate(
-                AppPage.CoinDetail("asdf")
-            )
-
-            is CoinListSideEffect.ShowToast -> {
-                Toast.makeText(context, effect.message, Toast.LENGTH_SHORT).show()
-            }
-
-            CoinListSideEffect.Close -> {
-                // 뒤로가기 로직 구현
-            }
-        }
-    }
-
-    CoinListScreen(
-        state = uiState,
-        onEvent = viewModel::handleEvent,
-        modifier = modifier,
-    )
-}
